Change the byte of the MemoryAttributes attribute to the size type
We prefer to use size_t type instead of byte type in PDI[1]

[1] https://gerrit.openbmc.org/c/openbmc/phosphor-dbus-interfaces/+/63799

+++ src/mdrv2.cpp
@@ -831,7 +831,7 @@ std::vector<boost::container::flat_map<std::string, RecordVariant>>
                                                    memoryInfo->length, dataIn);
             record["Part Number"] = positionToString(
                 memoryInfo->partNum, memoryInfo->length, dataIn);
-            record["Attributes"] = memoryInfo->attributes;
+            record["Attributes"] = uint32_t(memoryInfo->attributes);
             record["Extended Size"] = uint32_t(memoryInfo->extendedSize);
             record["Configured Memory Speed"] =
                 uint32_t(memoryInfo->confClockSpeed);
